// packages/sdk/src/generated/orval/models/likeMovieDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Match-movie example
 * The Match-Movie API description
 * OpenAPI spec version: 1.0
 */

export interface LikeMovieDto {
  /** Unique key of the room */
  roomKey: string;
  /** ID of the user */
  userId: number;
  /** ID of the movie */
  movieId: number;
}