// packages/sdk/src/generated/orval/models/moviesResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Match-movie example
 * The Match-Movie API description
 * OpenAPI spec version: 1.0
 */
import type { Movie } from './movie';

export interface MoviesResponse {
  /** Объекты фильмов */
  docs: Movie[];
  /** Общее количество фильмов */
  total: number;
  /** Лимит */
  limit: number;
  /** Текущая страница */
  page: number;
  /** Общее количество страниц */
  pages: number;
}