// packages/sdk/src/generated/orval/models/updateUserStatusDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Match-movie example
 * The Match-Movie API description
 * OpenAPI spec version: 1.0
 */
import type { UpdateUserStatusDtoUserStatus } from './updateUserStatusDtoUserStatus';

export interface UpdateUserStatusDto {
  /** Key of the room */
  roomKey: string;
  /** ID of the user */
  userId: number;
  /** Status of the user */
  userStatus: UpdateUserStatusDtoUserStatus;
}